refactor(server): extract sessions read handlers (#593)
Extract the 5 read handlers (api_sessions, api_sessions_local,
api_sessions_remote, api_worktrees, api_session_connections) from
server.py into agentwire/routes/sessions.py as SessionsRoutesMixin,
mirroring the scratchpad/machines pattern. Shared state and helpers
(active_sessions, _get_sessions_data, _fetch_remote_machine_sessions,
yaml read/write helpers, …) stay on the base server.

PR-A of 2 — read handlers. Mutate handlers ship in PR-B.

Part of #593

Built by dotdev.dev
